auth_service/requirements.txt
djangorestframework
python-decouple
django-cors-headers